requirements.txt
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
Pillow==10.0.0
PyYAML==6.0.1
tqdm==4.66.1
kaggle==1.5.16
opencv-python==4.8.0.76
jupyter==1.0.0
ipykernel==6.25.0
flask==3.0.0
gunicorn==21.2.0
